requirements.txt
annotated-types==0.6.0
anyio==3.7.1
attrs==23.1.0
certifi==2023.11.17
charset-normalizer==3.3.2
distro==1.8.0
exceptiongroup==1.2.0
fastjsonschema==2.19.0
filelock==3.13.1
fsspec==2023.10.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
huggingface-hub==0.19.4
idna==3.6
joblib==1.3.2
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
jupyter_core==5.5.0
llama-cpp-python==0.2.19
nbformat==5.9.2
numpy==1.26.2
openai==1.3.5
packaging==23.2
pandas==2.1.3
platformdirs==4.0.0
pydantic==2.5.2
pydantic_core==2.14.5
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.31.0
requests==2.31.0
rpds-py==0.13.1
scikit-learn==1.3.2
scipy==1.11.4
six==1.16.0
sniffio==1.3.0
threadpoolctl==3.2.0
tiktoken==0.5.1
tqdm==4.66.1
traitlets==5.13.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
